telecom/requirements.txt
# Core Django packages
Django==4.0.6
asgiref==3.5.2
daphne==3.0.2
channels==3.0.5

# Django extensions and plugins
django-bootstrap4==22.1
django-rest-framework==0.1.0 
djangorestframework==3.13.1
sqlparse==0.4.2

# Asynchronous libraries
aioredis==1.3.1
async-timeout==4.0.2
channels-redis==3.4.1
Twisted==22.4.0
txaio==22.2.1
typing-extensions==4.3.0

# Websocket and networking
autobahn==22.6.1
hyperlink==21.0.0
hiredis==2.0.0

# Web scraping and HTML parsing
beautifulsoup4==4.11.1
soupsieve==2.3.2.post1

# Utilities and helpers
Faker==13.15.1
factory-boy==3.2.1
idna==3.3
incremental==21.3.0
msgpack==1.0.4
PyYAML==6.0
service-identity==21.1.0
six==1.16.0
uritemplate==4.1.1
python-dateutil==2.8.2
pytz==2022.1

# Security and encryption
cffi==1.15.1
cryptography==37.0.4
pyOpenSSL==22.0.0
zope.interface==5.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8

# Image processing
Pillow==9.2.0

# Others..
attrs==21.4.0
Automat==20.2.0
constantly==15.1.0
pycparser==2.21

# For API documentation and testing
drf-yasg==1.20.0
coreapi==2.3.3

# For model utilities and mixins
django-model-utils==4.1.1

# For advanced searching and filtering in DRF
django-filter==2.4.0

# For authentication in DRF
djangorestframework-simplejwt==5.0.0

# For environmental variable management
python-decouple==3.5

# For managing static files
django-storages==1.11.1
boto3==1.17.112  # AWS S3 for static files
